aai_api/requirements.txt
Django>=5.2,<7
djangorestframework>=3.15,<4
Pillow>=10,<12
numpy>=1.26,<3
matplotlib
torch
torchvision

kaggle
certifi
grad-cam==1.5.5
opencv-python==4.13.0.92
python-dotenv==1.2.2
lime==0.2.0.1
scikit-image==0.26.0
captum==0.9.0
shap==0.51.0
scikit-learn
joblib
